move script to cli subpackage
